fix(toolchain): check for `package-lock.json` existence before modifying it

+++ scripts/sanitize-package-lock.js
@@ -2,10 +2,9 @@
 
 const fs = require('fs');
 
-fs.writeFileSync(
-	'package-lock.json',
-	fs.readFileSync('package-lock.json').toString().replace(/repo\.bit\.admin\.ch\/repository\/npm-group/g, 'registry.npmjs.org')
-);
-
-
-
+if (fs.existsSync('package-lock.json')) {
+	fs.writeFileSync(
+		'package-lock.json',
+		fs.readFileSync('package-lock.json').toString().replace(/repo\.bit\.admin\.ch\/repository\/npm-group/g, 'registry.npmjs.org')
+	);
+}
